Move and fix test
